Working on swapping out some of the nbt tags.
The chunk class should hold onto as few NBT tags as possible.
I'm not going to completely reverse engineer Minecraft to do
so, so some sections of the NBT will remain in the chunk like
the Entities tag because the data contained within is too
complex to separate into individual objects.